requirements.txt
mutagen==1.47.0
pillow==11.1.0
yt-dlp==2025.8.27
flask==3.0.0
flask-cors==4.0.0
requests==2.32.5
